test(core): Consolidate all execution repository unit tests in @n8n/db (no-changelog)
