+++ packages/@n8n/db/src/repositories/__tests__/execution.repository.test.ts
@@ -1,11 +1,18 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
+import { GlobalConfig } from '@n8n/config';
+import type { SqliteConfig } from '@n8n/config';
 import { Container } from '@n8n/di';
-import { In, LessThan, And, Not } from '@n8n/typeorm';
+import type { SelectQueryBuilder } from '@n8n/typeorm';
+import { In, LessThan, LessThanOrEqual, And, Not } from '@n8n/typeorm';
 import { mock } from 'jest-mock-extended';
+import { BinaryDataService } from 'n8n-core';
+import type { IRunExecutionData, IWorkflowBase } from 'n8n-workflow';
+import { nanoid } from 'nanoid';
 
 import { ExecutionEntity } from '../../entities';
 import type { IExecutionResponse } from '../../entities/types-db';
 import { mockEntityManager } from '../../utils/test-utils/mock-entity-manager';
+import { mockInstance } from '../../utils/test-utils/mock-instance';
 import { ExecutionRepository } from '../execution.repository';
 
 const GREATER_THAN_MAX_UPDATE_THRESHOLD = 901;
@@ -15,6 +22,10 @@ const GREATER_THAN_MAX_UPDATE_THRESHOLD = 901;
  */
 describe('ExecutionRepository', () => {
 	const entityManager = mockEntityManager(ExecutionEntity);
+	const globalConfig = mockInstance(GlobalConfig, {
+		logging: { outputs: ['console'], scopes: [] },
+	});
+	mockInstance(BinaryDataService);
 	const executionRepository = Container.get(ExecutionRepository);
 
 	beforeEach(() => {
@@ -495,5 +506,95 @@ describe('ExecutionRepository', () => {
 				expect.objectContaining({ status: 'canceled', waitTill: null }),
 			);
 		});
+	});
+
+	describe('getWaitingExecutions', () => {
+		const mockDate = new Date('2023-12-28 12:34:56.789Z');
+
+		beforeAll(() => jest.useFakeTimers().setSystemTime(mockDate));
+		afterAll(() => jest.useRealTimers());
+
+		test.each(['sqlite', 'postgresdb'] as const)(
+			'on %s, should only return executions with status=waiting',
+			async (dbType) => {
+				globalConfig.database.type = dbType;
+				entityManager.find.mockResolvedValueOnce([]);
+
+				await executionRepository.getWaitingExecutions();
+
+				expect(entityManager.find).toHaveBeenCalledWith(ExecutionEntity, {
+					order: { waitTill: 'ASC' },
+					select: ['id', 'waitTill'],
+					where: {
+						status: 'waiting',
+						waitTill: LessThanOrEqual(
+							dbType === 'sqlite'
+								? '2023-12-28 12:36:06.789'
+								: new Date('2023-12-28T12:36:06.789Z'),
+						),
+					},
+				});
+			},
+		);
+	});
+
+	describe('deleteExecutionsByFilter', () => {
+		test('should delete binary data', async () => {
+			const workflowId = nanoid();
+			const binaryDataService = Container.get(BinaryDataService);
+
+			jest.spyOn(executionRepository, 'createQueryBuilder').mockReturnValue(
+				mock<SelectQueryBuilder<ExecutionEntity>>({
+					select: jest.fn().mockReturnThis(),
+					andWhere: jest.fn().mockReturnThis(),
+					getMany: jest.fn().mockResolvedValue([{ id: '1', workflowId }]),
+				}),
+			);
+
+			await executionRepository.deleteExecutionsByFilter({
+				filters: { id: '1' },
+				accessibleWorkflowIds: ['1'],
+				deleteConditions: { ids: ['1'] },
+			});
+
+			expect(binaryDataService.deleteMany).toHaveBeenCalledWith([
+				{ type: 'execution', executionId: '1', workflowId },
+			]);
+		});
+	});
+
+	describe('updateExistingExecution', () => {
+		test.each(['sqlite', 'postgresdb'] as const)(
+			'should update execution and data in transaction on %s',
+			async (dbType) => {
+				globalConfig.database.type = dbType;
+				globalConfig.database.sqlite = mock<SqliteConfig>({ poolSize: 1 });
+
+				const executionId = '1';
+				const execution = mock<IExecutionResponse>({
+					id: executionId,
+					data: mock<IRunExecutionData>(),
+					workflowData: mock<IWorkflowBase>(),
+					status: 'success',
+				});
+
+				const txCallback = jest.fn();
+				entityManager.transaction.mockImplementation(async (fn: unknown) => {
+					await (fn as (em: typeof entityManager) => Promise<unknown>)(entityManager);
+					txCallback();
+				});
+				entityManager.update.mockResolvedValue({ affected: 1, raw: [], generatedMaps: [] });
+
+				await executionRepository.updateExistingExecution(executionId, execution);
+
+				expect(entityManager.transaction).toHaveBeenCalled();
+				expect(entityManager.update).toHaveBeenCalledWith(
+					ExecutionEntity,
+					{ id: executionId },
+					expect.objectContaining({ status: 'success' }),
+				);
+				expect(txCallback).toHaveBeenCalledTimes(1);
+			},
+		);
 	});
 });
